Rolled back to 1347

+++ tests/test-finit.C
@@ -29,7 +29,6 @@
 
 // #define SIMD_INT
 
-#include "fflas-ffpack/fflas-ffpack-config.h"
 #include <typeinfo>
 #include <givaro/modular.h>
 #include <givaro/modular-balanced.h>
@@ -186,11 +185,11 @@ int main(int ac, char **av) {
 			pass &= test_freduce (F,m,k,n,timing);
 		}
 		{
-			Givaro::Modular<int32_t> F((int32_t)p) ;
+			Givaro::Modular<int32_t> F(p) ;
 			pass &= test_freduce (F,m,k,n,timing);
 		}
 		{
-			Givaro::ModularBalanced<int32_t> F((int32_t)p) ;
+			Givaro::ModularBalanced<int32_t> F((int)p) ;
 			pass &= test_freduce (F,m,k,n,timing);
 		}
 		{
